+++ src/lib/forms-schema/element-schema.ts
@@ -37,6 +37,7 @@ import FileElement from './elements/FileElement'
 //Integration
 import GeoscapeElement from './elements/GeoscapeElement'
 import PointElement from './elements/PointElement'
+import civicaStreetNameElement from './elements/CivicaStreetNameElement'
 
 const typeCase = (type: FormTypes.FormElementType) => {
   return Joi.object({
@@ -142,6 +143,9 @@ const schema = Joi.object({
   })
   .when(typeCase('boolean'), {
     then: BooleanElement,
+  })
+  .when(typeCase('civicaStreetName'), {
+    then: civicaStreetNameElement,
   })
   .id('formElement')