+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Boolean/Switch element
+- `Boolean` form element
+- `CIVICA_CRM` form submission event
+- `civicaStreetName` form element
 
 ## [0.9.14] - 2021-06-06